+++ tests/test_base.py
@@ -4,9 +4,8 @@ from mirakuru import *  # noqa
 
 def test_importing_mirakuru():
     """Test if all most commonly used classes are imported by default."""
-    assert HTTPExecutor
-    assert TCPExecutor
     assert Executor
+    assert SimpleExecutor
     assert OutputExecutor
     assert TCPExecutor
     assert HTTPExecutor